read-config.go -> générer un capteur a partir d'un fichier de config
